pyproject.toml
[project]
name = "kafka-handson-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.28",
    "kafka-python>=2.2.10",
    "requests>=2.32.3",
]
